move magnetic_pressure to plasmapy.formulary.misc
