+++ plasmapy/formulary/parameters.py
@@ -1,14 +1,10 @@
 """Functions to calculate fundamental plasma parameters."""
 
-__all__ = ["magnetic_pressure"]
-__aliases__ = ["pmag_"]
-__lite_funcs__ = []
+__all__ = []
 
-import astropy.units as u
+from astropy.constants.si import e, eps0, k_B
 
-from astropy.constants.si import e, eps0, k_B, mu0
-
-from plasmapy.utils.decorators import deprecated, validate_quantities
+from plasmapy.utils.decorators import deprecated
 from plasmapy.utils.exceptions import PlasmaPyFutureWarning
 
 from plasmapy.formulary import dimensionless, frequencies, lengths, misc, speeds  # noqa
@@ -20,11 +16,9 @@ __all__ += (
     + lengths.__all__.copy()
     + misc.__all__.copy()
     + speeds.__all__.copy()
-    + __aliases__
-    + __lite_funcs__
 )
 
-__aliases__ += (
+__aliases__ = (
     dimensionless.__aliases__.copy()
     + frequencies.__aliases__.copy()
     + lengths.__aliases__.copy()
@@ -32,7 +26,7 @@ __aliases__ += (
     + speeds.__aliases__.copy()
 )
 
-__lite_funcs__ += frequencies.__lite_funcs__.copy() + speeds.__lite_funcs__.copy()
+__lite_funcs__ = frequencies.__lite_funcs__.copy() + speeds.__lite_funcs__.copy()
 
 # remove from __all__ and __aliases__ added functionality names that are not
 # actually in this file
@@ -88,6 +82,8 @@ funcs_to_deprecate_wrap = [  # (module_name, func_name)
     ("misc", "DB_"),
     ("misc", "magnetic_energy_density"),
     ("misc", "ub_"),
+    ("misc", "magnetic_pressure"),
+    ("misc", "pmag_"),
     ("misc", "mass_density"),
     ("misc", "rho_"),
     ("misc", "thermal_pressure"),
@@ -116,68 +112,3 @@ for modname, name in funcs_to_deprecate_wrap:
     )(getattr(globals()[f"{modname}"], name))
 
 del funcs_to_deprecate_wrap, modname, name
-
-
-@validate_quantities
-def magnetic_pressure(B: u.T) -> u.Pa:
-    r"""
-    Calculate the magnetic pressure.
-
-    **Aliases:** `pmag_`
-
-    Parameters
-    ----------
-    B : `~astropy.units.Quantity`
-        The magnetic field in units convertible to tesla.
-
-    Returns
-    -------
-    p_B : `~astropy.units.Quantity`
-        The magnetic pressure in units in pascals (newtons per square meter).
-
-    Raises
-    ------
-    `TypeError`
-        If the input is not a `~astropy.units.Quantity`.
-
-    `~astropy.units.UnitConversionError`
-        If the input is not in units convertible to tesla.
-
-    `ValueError`
-        If the magnetic field strength is not a real number between
-        :math:`±∞`\ .
-
-    Warns
-    -----
-    : `~astropy.units.UnitsWarning`
-        If units are not provided, SI units are assumed.
-
-    Notes
-    -----
-    The magnetic pressure is given by:
-
-    .. math::
-        p_B = \frac{B^2}{2 μ_0}
-
-    The motivation behind having two separate functions for magnetic
-    pressure and magnetic energy density is that it allows greater
-    insight into the physics that are being considered by the user and
-    thus more readable code.
-
-    See Also
-    --------
-    magnetic_energy_density : returns an equivalent `~astropy.units.Quantity`,
-        except in units of joules per cubic meter.
-
-    Examples
-    --------
-    >>> from astropy import units as u
-    >>> magnetic_pressure(0.1*u.T).to(u.Pa)
-    <Quantity 3978.87... Pa>
-
-    """
-    return (B ** 2) / (2 * mu0)
-
-
-pmag_ = magnetic_pressure
-"""Alias to `~plasmapy.formulary.parameters.magnetic_pressure`."""
